ai-poc/requirements.txt
fastapi
uvicorn[standard]
pillow
pytesseract
pdfplumber
spacy
python-dotenv
transformers
requests
sentence-transformers
faiss-cpu
numpy
httpx

